requirements.txt
pycryptodome
scapy
argparse
pathlib

